PFC-2444 #comment done

+++ src/main/de/dal33t/powerfolder/Controller.java
@@ -120,6 +120,7 @@ import de.dal33t.powerfolder.util.ProUtil;
 import de.dal33t.powerfolder.util.Profiling;
 import de.dal33t.powerfolder.util.PropertiesUtil;
 import de.dal33t.powerfolder.util.Reject;
+import de.dal33t.powerfolder.util.SplitConfig;
 import de.dal33t.powerfolder.util.StringUtils;
 import de.dal33t.powerfolder.util.Translation;
 import de.dal33t.powerfolder.util.Util;
@@ -167,9 +168,10 @@ public class Controller extends PFComponent {
      * The actual config file.
      */
     private Path configFile;
+    private Path configFolderFile;
 
     /** The config properties */
-    private Properties config;
+    private SplitConfig config;
 
     /**
      * The preferences
@@ -886,14 +888,13 @@ public class Controller extends PFComponent {
         }
 
         configFilename = null;
-        config = new Properties();
+        config = new SplitConfig();
         configFilename = filename;
         configFile = getConfigLocationBase();
-
+        
         if (configFile == null) {
             configFile = Paths.get(filename).toAbsolutePath();
-        }
-        else {
+        } else {
             configFile = configFile.resolve(filename);
         }
 
@@ -927,6 +928,43 @@ public class Controller extends PFComponent {
             } catch (Exception e) {
                 // Ignore.
             }
+        }
+        
+        String folderfilename = filename.replace(".config", "-Folder.config");
+        configFolderFile = getConfigLocationBase();
+        if (configFolderFile == null) {
+            configFolderFile = Paths.get(folderfilename).toAbsolutePath();
+        } else {
+            configFolderFile = configFolderFile.resolve(folderfilename);
+        }
+
+        if (Files.exists(configFolderFile)) {
+            try {
+                logInfo("Loading folder configfile "
+                    + configFolderFile.toString());
+                bis = new BufferedInputStream(
+                    Files.newInputStream(configFolderFile));
+                config.load(bis);
+            } catch (FileNotFoundException e) {
+                logWarning("Unable to start config, file '" + folderfilename
+                    + "' not found, using defaults");
+            } catch (IOException e) {
+                logSevere("Unable to start config from file '" + folderfilename
+                    + '\'');
+                configFolderFile = null;
+                return false;
+            } finally {
+                try {
+                    if (bis != null) {
+                        bis.close();
+                    }
+                } catch (Exception e) {
+                    // Ignore.
+                }
+            }
+        } else {
+            logFine("Folder config file does not exist. "
+                + configFolderFile.toString());
         }
         return true;
     }
@@ -1224,6 +1262,16 @@ public class Controller extends PFComponent {
         } catch (IOException e) {
             logWarning("Unable to backup file " + configFile + ". " + e);
         }
+        if (Files.exists(configFolderFile)) {
+            Path configFolderBackup = backupDir.resolve(configFolderFile
+                .getFileName());
+            try {
+                PathUtils.copyFile(configFolderFile, configFolderBackup);
+            } catch (IOException e) {
+                logWarning("Unable to backup file " + configFolderFile + ". "
+                    + e);
+            }
+        }
         Path myKeyFile = getMiscFilesLocation().resolve(getConfigName()
             + ".mykey");
         Path mykeyBackup = backupDir.resolve(myKeyFile.getFileName());
@@ -1428,11 +1476,17 @@ public class Controller extends PFComponent {
 
         Path file;
         Path tempFile;
+        Path folderFile;
+        Path tempFolderFile;
         Path backupFile;
         if (getConfigLocationBase() == null) {
             file = Paths.get(getConfigName() + ".config").toAbsolutePath();
             tempFile = Paths.get(getConfigName() + ".writing.config")
                 .toAbsolutePath();
+            folderFile = Paths.get(getConfigName() + "-Folder.config")
+                .toAbsolutePath();
+            tempFolderFile = Paths.get(
+                getConfigName() + "-Folder.writing.config").toAbsolutePath();
             backupFile = Paths.get(getConfigName() + ".config.backup")
                 .toAbsolutePath();
         } else {
@@ -1441,6 +1495,10 @@ public class Controller extends PFComponent {
                 getConfigName() + ".writing.config").toAbsolutePath();
             backupFile = getConfigLocationBase().resolve(
                 getConfigName() + ".config.backup");
+            folderFile = getConfigLocationBase().resolve(
+                getConfigName() + "-Folder.config");
+            tempFolderFile = getConfigLocationBase().resolve(
+                getConfigName() + "-Folder.writing.config").toAbsolutePath();
         }
 
         try {
@@ -1453,19 +1511,44 @@ public class Controller extends PFComponent {
             {
                 distName = distribution.getName();
             }
-            // Store config in misc base
-            PropertiesUtil.saveConfig(tempFile, config, distName
-                + " config file (v" + PROGRAM_VERSION + ')');
-            
 
-            Files.deleteIfExists(file);
-            try {
-                Files.move(tempFile, file);
-            } catch (IOException e) {
-                Files.copy(tempFile, file);
-                Files.delete(tempFile);
+            Properties prev = new Properties();
+            if (Files.exists(file)) {
+                try (BufferedInputStream in = new BufferedInputStream(
+                    Files.newInputStream(file))) {
+                    prev.load(in);
+                }
             }
 
+            if (!prev.equals(config.getRegular())) {
+                // Store config in misc base
+                PropertiesUtil.saveConfig(tempFile, config.getRegular(),
+                    distName + " config file (v" + PROGRAM_VERSION + ')');
+                Files.deleteIfExists(file);
+                try {
+                    Files.move(tempFile, file);
+                } catch (IOException e) {
+                    Files.copy(tempFile, file);
+                    Files.delete(tempFile);
+                }
+            } else {
+                if (isFine()) {
+                    logFine("Not storing config to " + file + ". Base config remains unchanged");                    
+                }
+            }
+
+            if (!config.getFolders().isEmpty()) {
+                PropertiesUtil.saveConfig(tempFolderFile, config.getFolders(),
+                    distName + " folders config file (v" + PROGRAM_VERSION
+                        + ')');
+                Files.deleteIfExists(folderFile);
+                try {
+                    Files.move(tempFolderFile, folderFile);
+                } catch (IOException e) {
+                    Files.copy(tempFolderFile, folderFile);
+                    Files.delete(tempFolderFile);
+                }
+            }
         } catch (IOException e) {
             // FATAL
             logSevere("Unable to save config. " + e, e);
@@ -1948,6 +2031,10 @@ public class Controller extends PFComponent {
 
     public Path getConfigFile() {
         return configFile;
+    }
+    
+    public Path getConfigFolderFile() {
+        return configFolderFile;
     }
 
     /**
